+++ src/main/java/lotto/controller/BonusNumberController.java
@@ -1,9 +1,9 @@
 package lotto.controller;
 
 import lotto.model.BonusNumber;
-import lotto.temp.IoComponent;
+import lotto.util.IoComponent;
 
-import static lotto.util.RepeatInput.repeatUntilValid;
+import static lotto.util.common.RepeatInput.repeatUntilValid;
 
 public class BonusNumberController {
     private final IoComponent ioComponent;
