+++ src/main/java/lotto/util/common/CommonIo.java
@@ -1,4 +1,4 @@
-package lotto.util;
+package lotto.util.common;
 
 import camp.nextstep.edu.missionutils.Console;
