+++ src/test/java/lotto/model/StaticsTest.java
@@ -1,4 +1,4 @@
-package lotto.temp;
+package lotto.model;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
